Made the record editing code more robust by waiting for input for a few seconds, looking for a string to indicate status, before trying to create the temp file.  Also added the ability to edit dictionary records by adding `DICT' on the edit command, or as a parameter when calling the emacs functions directly.
Compiling and cataloging work now, because all unidata connections are
kept in a list, and it looks in that list for a connection to use.  Now,
it simply picks the first one (which works for me now because I only run
one connection at a time anyway); I should add a search for connections
where a unidata file is likely to be able to be compiled.

One can also run unibasic programs from within the unibasic buffer, as
long as they take no arguments.  This should be improved.